[whisper] move processor test into processor test file 🧹  (#38266)
move processor tests
